Плеер подключён к хранилищу с помощью хуков
