--- project/src/components/_screen/player-screen/player-screen-style.ts
@@ -1,12 +1,0 @@
-import dayjs from 'dayjs';
-import duration from 'dayjs/plugin/duration';
-dayjs.extend(duration);
-
-const SECONDS_IN_MINUTES = 60;
-
-const getTimeFormat = (currentTime: number): string => {
-  const timeFormat = currentTime >= (SECONDS_IN_MINUTES * SECONDS_IN_MINUTES) ? '-HH:mm:ss' : '-mm:ss';
-  return dayjs.duration(currentTime, 'seconds').format(timeFormat);
-};
-
-export {getTimeFormat};
