+++ project/src/store/film-data/selectors.ts
@@ -5,7 +5,7 @@ import {getFilmsWithoutId} from '../../utils/utils';
 import {Comments} from '../../types/comment';
 
 const getCurrentFilm = (state: State): Film | null => (state[NameSpace.film].currentFilm);
-const getCurrentFilmById = (state: State, id: number): Film | undefined => (state[NameSpace.films].films.find((item) => item.id === id));
+const getCurrentFilmById = (id: number) => (state: State): Film | undefined => (state[NameSpace.films].films.find((item) => item.id === id));
 const getLoadingStatus = (state: State): boolean => (state[NameSpace.film].isLoading);
 const getComments = (state: State): Comments => (state[NameSpace.film].comments);
 const getFavoriteFilms = (state: State): Films => (state[NameSpace.film].favoriteFilms);
